+++ cli/cli/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `--dotnet-path` option is available to override dotnet installation location.
+- Auto generation handles trials and session services.
 
 ## [1.19.2]